backend/requirements.txt
pdfplumber==0.10.3
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
flask-cors==4.0.0
redis==7.4.0
APScheduler==3.10.4
